feat(address): adding dependencies to all address routes

+++ app/dao/dao_address.py
@@ -1,15 +1,15 @@
 from app.dao.dao import connect_database
 from app.schemas.address import Address
+from app.dao.dao_employee import delete_employees
 
-
-def select_address(address_id: int):
+def select_address(employee_id: int):
     
     connection, cursor = connect_database()
     
     query = f"""
-    SELECT *
-    from Address c 
-    WHERE id = {address_id}
+    SELECT a.num, a.complement, a.zipcode, a.street, a.district, a.city, a.state  FROM onboarding_me.Address a
+    LEFT JOIN Employee e ON e.address_id = a.id
+    WHERE e.id = {employee_id}
     ;
     """
 
@@ -25,6 +25,35 @@ def select_address(address_id: int):
         connection.close()
         
         return address
+    
+    
+def select_employees_by_address(address_id: int):
+    
+    connection, cursor = connect_database()
+    
+    query = f"""
+    SELECT  e.id FROM onboarding_me.Address a 
+    LEFT JOIN Employee e ON e.address_id = a.id
+    WHERE a.id = {address_id}
+    ;
+    """
+    
+    try:
+        cursor.execute(query)
+        
+    except Exception as error:
+        connection.close()
+        return None
+    else:
+        employees_list = cursor.fetchall()
+        connection.close()
+        
+        ids_list = []
+        
+        for employee in employees_list:
+            ids_list.append(employee['id'])
+        
+        return ids_list     
     
 
 def insert_address(address: Address):
@@ -77,11 +106,19 @@ def update_address(address: Address):
         connection.commit()
         connection.close()
         return True
-    
+
     
 def delete_address(address_id: int):
     
     connection, cursor = connect_database()
+    
+    employee_list = select_employees_by_address(address_id)
+    
+    employee_deleted = delete_employees(employee_list)
+    
+    if not employee_deleted:
+        return False
+    
     
     query = f"""
     DELETE FROM onboarding_me.Address
@@ -92,7 +129,7 @@ def delete_address(address_id: int):
     try:
         cursor.execute(query)
         
-    except:
+    except Exception as error:
         connection.close()
         return False
     else:
@@ -102,16 +139,27 @@ def delete_address(address_id: int):
         return True    
     
     
-def verify_if_address_exists_by_id(address_id: int):
+def verify_if_address_exists(employee_id: int = None, address_id: int = None):
     
     connection, cursor = connect_database()
     
-    query = f"""
-    SELECT id
-    FROM onboarding_me.Address
-    WHERE id = {address_id}
-    ;
-    """
+    if employee_id:
+        
+        query = f"""
+        SELECT a.num FROM Address a 
+        LEFT JOIN Employee e ON e.address_id = a.id
+        WHERE e.id = {employee_id}
+        ;
+        """
+    
+    if address_id:
+        
+        query = f"""
+        SELECT id
+        FROM onboarding_me.Address
+        WHERE id = {address_id}
+        ;
+        """
     
     try:
         cursor.execute(query)
@@ -131,14 +179,43 @@ def verify_if_address_exists_by_id(address_id: int):
     return False
 
 
-def verify_if_address_exists_by_house(num: str, street: str):
+def return_company_addresses(company_id: NotImplemented):
+    
+    connection, cursor = connect_database()
+        
+    query = f"""
+    SELECT DISTINCT a.id  FROM onboarding_me.Address a 
+    LEFT JOIN Employee e ON e.address_id = a.id
+    LEFT JOIN Company c ON c.id = e.company_id 
+    WHERE c.id = {company_id}
+    ;
+    """
+
+    
+    try:
+        cursor.execute(query)
+        
+    except Exception as error:
+        connection.close()
+        return None    
+    
+    else:    
+
+        addresses_list = cursor.fetchall()
+        connection.close()
+        
+        return addresses_list
+
+
+def verify_if_address_exists_by_house(company_id: int):
     
     connection, cursor = connect_database()
     
     query = f"""
-    SELECT num
-    FROM onboarding_me.Address
-    WHERE num = '{num}' AND street = '{street}'
+    SELECT a.num, a.street  FROM Company c
+    RIGHT JOIN Employee e ON c.id = e.company_id 
+    RIGHT JOIN Address a ON e.address_id = a.id 
+    WHERE c.id = {company_id}
     ;
     """
     
@@ -147,14 +224,11 @@ def verify_if_address_exists_by_house(num: str, street: str):
         
     except Exception as error:
         connection.close()
-        return False    
+        return None    
     
     else:    
         
-        address_exists = cursor.fetchone()
+        addresses_list  = cursor.fetchall()
         connection.close()
         
-        if address_exists:
-            return True
-        
-    return False
+        return addresses_list
